Store signals the script accepts.
It is dangerous to access the lua interpreter from another thread than
the thread running the lua code. If the lua thread is busy this leads to
locks (that can lead to deadlocks). To prevent this we store the
accepted signals in a list so we don't need to access the lua global
object from our main thread.