Add extra blockheater limit
If a trigger time has been set, but the temperature goes above
10 degrees within two hours before departure, we are still triggering,
since the engine itself might still be cold even if the sun has
come up and is warming the frozen souls up in the north. However,
if the temperature goes above 13 degrees, just stop the trigger.
That would lead to an negative offset in any case.