Convert python3 .keys()/.values() to lists as before
Dictionaries iteration methods returns views in python3, instead
of lists as before. In some cases it might not have been necessary
to convert them to lists, when this was guaranteed they were not
converted to lists, even if suggested by the 2to3-tool.

The situation is similar with "range", which now is a range-
object, instead of list as before. But I could find no places
where range could not be used instead, so these where not
converted.