Unify how failed state changes are reported to the Live! server
To make this work correctly, the servers should also be upgraded
(when new firmware has been pushed) to not log failed commands
to the device history when nack is received (only remove from
ack list then), and instead wait for a devicefailevent.

If servers are not upgraded, failed state changes inititated
server side will be logged twice.